cliente/requirements.txt
art==5.6
python_dateutil==2.8.2
requests==2.22.0